Add password input component and put it in node operator widgets
