+++ src/react/components/DecryptValidatorKeysWidget.tsx
@@ -1,22 +1,19 @@
 import React, { useState } from 'react';
-import {
-    Box, Center, Button, VStack, Text,
-    HStack, NumberInput, NumberInputField,
-    InputRightElement, InputGroup,
-} from '@chakra-ui/react';
+import { Box, Center, Button, VStack, Text, HStack } from '@chakra-ui/react';
 import raisedWidgetStyle from '../styleClasses/widgetBoxStyle';
-import successBoxStyle from '../styleClasses/successBoxStyle';
 import darkBoxWithBorderStyle from '../styleClasses/darkBoxWithBorderStyle';
 import SelectFile from './SelectFile'
+import PasswordInput from './PasswordInput';
 
 
 
 const DecryptValidatorKeysWidget: React.FC = () => {
-    const [numKeys, setNumKeys] = useState<string>("500");
     const [savePath, setSavePath] = useState<string>("")
     const [keysGenerated, setKeysGenerated] = useState<boolean>(false)
     const [encryptedValKeysFilePath, setencryptedValidatorKeysFilePath] = useState<string>("")
     const [privKeysFilePath, setPrivKeysFilePath] = useState<string>("")
+    const [privKeysPassword, setPrivKeysPassword] = useState<string>("")
+    const [isPrivKeysPasswordValid, setIsPrivKeysPasswordValid] = useState<boolean>(false)
 
     const selectSavePath = () => {
         window.api.receiveSelectedFolderPath((event: Electron.IpcMainEvent, path: string) => {
@@ -30,7 +27,7 @@ const DecryptValidatorKeysWidget: React.FC = () => {
             console.log("DECRYPT COMPLETE!")
             console.log(path)
         })
-        window.api.reqDecryptValidatorKeys(encryptedValKeysFilePath, privKeysFilePath, savePath);
+        window.api.reqDecryptValidatorKeys(encryptedValKeysFilePath, privKeysFilePath, privKeysPassword, savePath);
     }
 
 
@@ -61,6 +58,10 @@ const DecryptValidatorKeysWidget: React.FC = () => {
                                         <SelectFile setFilePath={setPrivKeysFilePath} filePath={privKeysFilePath} />
                                     </Box>
                                     <Box>
+                                        <PasswordInput password={privKeysPassword} setPassword={setPrivKeysPassword} isPasswordValid={isPrivKeysPasswordValid} setIsPasswordValid={setIsPrivKeysPasswordValid} />
+
+                                    </Box>
+                                    <Box>
                                         <Text color="white" opacity={'0.7'} align="center">
                                             {savePath}
                                         </Text>
@@ -71,7 +72,7 @@ const DecryptValidatorKeysWidget: React.FC = () => {
                                 <Center>
                                     <HStack spacing='10px' mb="5px">
                                         <Button variant="white-button" onClick={selectSavePath}>{savePath ? "Change Path" : "Select Save Path"}</Button>
-                                        <Button variant="white-button" isDisabled={!savePath || !encryptedValKeysFilePath || !privKeysFilePath}
+                                        <Button variant="white-button" isDisabled={!savePath || !encryptedValKeysFilePath || !privKeysFilePath || !isPrivKeysPasswordValid}
                                             _disabled={{ bg: "grey.dark", _hover: { bg: "grey.dark" } }} onClick={decryptValidatorKeys}>
                                             Decrypt Validator Keys
                                         </Button>
@@ -79,11 +80,11 @@ const DecryptValidatorKeysWidget: React.FC = () => {
                                 </Center>
                             </Box>
                         </VStack>
-                    </Box>
+                    </Box >
 
                 )
             }
-        </Center>
+        </Center >
     )
 }
